not working 10

+++ lib/array_d.c
@@ -1,18 +1,19 @@
 //
-// Created by Mateusz Chojnowski on 13.04.2021.
+// Created by mateu on 11.05.2021.
 //
-#include "array.h"
+
+#include "array_d.h"
 #include <assert.h>
 #include <stdio.h>
 
 array create_array(size_t size) {
-    float *data = malloc(size * sizeof(float));
+    double *data = malloc(size * sizeof(double));
     array arr = {size, data};
     return arr;
 }
 
 array create_zero_array(size_t size) {
-    float *data = calloc(size, sizeof(float));
+    double *data = calloc(size, sizeof(double));
     array arr = {size, data};
     return arr;
 }
@@ -20,19 +21,19 @@ array create_zero_array(size_t size) {
 void move_array(array *from, array *to) {
     free(to->data);
     *((size_t*) &to->size) = from->size;
-    *((float **) &to->data) = from->data;
+    *((double **) &to->data) = from->data;
     *((size_t*) &from->size) = -1;
-    *((float**) &from->data) = NULL;
+    *((double**) &from->data) = NULL;
 }
 
 void zero_array(array arr) {
     for (size_t i = 0; i < arr.size; ++i)
-        arr.data[i] = 0.0f;
+        arr.data[i] = 0.0;
 }
 
-float arrdot(array first, array second) {
+double arrdot(array first, array second) {
     assert(first.size == second.size);
-    float sum = 0.0f;
+    double sum = 0.0;
     for (int i = 0; i < first.size; ++i) {
         sum += first.data[i] * second.data[i];
     }
@@ -43,12 +44,12 @@ void destroy_array(array arr) {
     if (arr.data != NULL)
         free(arr.data);
     *((size_t*) &arr.size) = -1;
-    *((float**) &arr.data) = NULL;
+    *((double**) &arr.data) = NULL;
 }
 
 void print_array(array arr) {
     for (int i = 0; i < arr.size; ++i) {
-        printf("%f ", arr.data[i]);
+        printf("%lf ", arr.data[i]);
     }
     printf("\n");
 }
